+++ UHDD/src/application/CredentialManager.java
@@ -9,7 +9,7 @@ public class CredentialManager {
 	DBConnector dbConnector = new DBConnector();
 	PasswordHasher passwordHasher = new PasswordHasher();
 	
-	protected void addNewUserToDB(String username, String password, String email, String role) throws ClassNotFoundException, SQLException {
+	public void addNewUserToDB(String username, String password, String email, String role) throws ClassNotFoundException, SQLException {
 		dbConnector.initialiseDB();		
 		PasswordHash passwordHash = passwordHasher.hashPassword(password);
 		String passwordHashAsString = passwordHash.getHashAsString();		
@@ -19,7 +19,7 @@ public class CredentialManager {
 	}
 	
 	
-	protected String checkCredentialsInFile(String username, String password) throws ClassNotFoundException, SQLException {
+	public String checkCredentialsInFile(String username, String password) throws ClassNotFoundException, SQLException {
 		dbConnector.initialiseDB();
 	    
 		try (ResultSet userDetails = dbConnector.QueryReturnResultsFromUser(username)) {
