+++ UHDD/src/application/fxmlScenes/DoctorView.fxml
@@ -50,7 +50,7 @@
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<VBox prefHeight="838.0" prefWidth="1197.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.UI">
+<VBox prefHeight="838.0" prefWidth="1197.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.viewControllers.UI">
   <children>
     <MenuBar prefWidth="1141.0" style="-fx-background-color: #527c88;" VBox.vgrow="NEVER">
       <menus>
