+++ UHDD/src/application/fxmlScenes/Nurse_Dashboard.fxml
@@ -44,7 +44,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<AnchorPane maxHeight="-1.0" maxWidth="-1.0" prefHeight="506.0" prefWidth="916.0" style="-fx-background-color: #fcfcfc;" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.Nurse_DashboardController">
+<AnchorPane maxHeight="-1.0" maxWidth="-1.0" prefHeight="506.0" prefWidth="916.0" style="-fx-background-color: #fcfcfc;" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.viewControllers.Nurse_DashboardController">
 <children>
       <Pane prefHeight="506.0" prefWidth="200.0" style="-fx-background-color: #4B0082;">
          <children>
