+++ UHDD/src/application/fxmlScenes/PatientDirectoryOLD.fxml
@@ -47,7 +47,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<VBox prefHeight="600.0" prefWidth="900.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.PatientDirectoryController">
+<VBox prefHeight="600.0" prefWidth="900.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.viewControllers.PatientDirectoryController">
   <children>
     <MenuBar minWidth="1500.0" VBox.vgrow="NEVER">
       <menus>
