+++ UHDD/src/application/fxmlScenes/PatientInfoView2.fxml
@@ -38,7 +38,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<AnchorPane maxHeight="-1.0" maxWidth="-1.0" prefHeight="506.0" prefWidth="916.0" style="-fx-background-color: #e0e0e0;" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.PatientInfoViewController">
+<AnchorPane maxHeight="-1.0" maxWidth="-1.0" prefHeight="506.0" prefWidth="916.0" style="-fx-background-color: #e0e0e0;" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.viewControllers.PatientInfoViewController">
    <children>
       <Pane prefHeight="506.0" prefWidth="200.0" style="-fx-background-color: #063847;">
          <children>
