+++ UHDD/src/application/fxmlScenes/UserCreation.fxml
@@ -13,7 +13,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="506.0" prefWidth="916.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.UserCreationController">
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="506.0" prefWidth="916.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.viewControllers.UserCreationController">
    <children>
       <ImageView fx:id="imgeViewCreateUser" fitHeight="506.0" fitWidth="916.0" layoutX="79.0" layoutY="-1.0" opacity="0.3" pickOnBounds="true" preserveRatio="true" scaleX="1.25">
          <image>
