+++ UHDD/src/application/viewControllers/Nurse_PatientDirectoryController.java
@@ -1,20 +1,23 @@
-package application;
+package application.viewControllers;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
+import application.DBConnector;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.value.ObservableValue;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
+import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Node;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.TableColumn.CellDataFeatures;
+import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.Pane;
 import javafx.scene.control.Button;
@@ -23,7 +26,7 @@ import javafx.scene.control.TableView;
 import javafx.stage.Stage;
 import javafx.util.Callback;
 
-public class PatientDirectoryController {
+public class Nurse_PatientDirectoryController {
 	private Stage stage;
 	private Scene scene;
 	@FXML private Button back;
@@ -44,7 +47,7 @@ public class PatientDirectoryController {
 	DBConnector dbConnection = new DBConnector();
 	
 	public void switchToDashBoard(MouseEvent mouseEvent) throws IOException {
-		Parent root = FXMLLoader.load(getClass().getResource("\\fxmlScenes\\DashBoard.fxml"));
+		Parent root = FXMLLoader.load(getClass().getResource("\\fxmlScenes\\Nurse_DashBoard.fxml"));
 		stage = (Stage)((Node)mouseEvent.getSource()).getScene().getWindow();
 		scene = new Scene(root);
 		stage.setScene(scene);
@@ -89,7 +92,7 @@ public class PatientDirectoryController {
                 for (int i = 1; i <= rs.getMetaData().getColumnCount(); i++) {
                     //Iterate Column
                     row.add(rs.getString(i));
-                    System.out.println("Counting column...");
+                    System.out.println("Counting column..." + i);
                 }
                 System.out.println("Row [1] added " + row);
                 data.add(row);
