+++ UHDD/src/application/viewControllers/Nurse_PatientInfoViewController.java
@@ -1,10 +1,11 @@
-package application;
+package application.viewControllers;
 
 import java.io.IOException;
 import java.sql.ResultSet;
 import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
 
+import application.DBConnector;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Node;
