+++ UHDD/src/application/viewControllers/Nurse_PatientInformationController.java
@@ -1,4 +1,4 @@
-package application;
+package application.viewControllers;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
@@ -7,6 +7,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.time.LocalDate;
 
+import application.DBConnector;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
