+++ UHDD/src/application/viewControllers/TableCreatorController.java
@@ -1,4 +1,4 @@
-package application;
+package application.viewControllers;
 
 import java.io.File;
 import java.io.FileNotFoundException;
@@ -7,6 +7,8 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
 
+import application.DBConnector;
+import application.MySQL_test;
 import javafx.beans.property.ReadOnlyStringWrapper;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.value.ObservableValue;
